ci update licence prohibition rules

+++ src/actions/components/connector/src/TDengineConnector.cpp
@@ -292,7 +292,7 @@ bool TDengineConnector::execute(const StmtV2InsertData& data) {
             "{} failed to bind parameters: {} [code: 0x{:x}]",
             display_name_,
             taos_stmt2_error_(stmt_),
-            code
+            static_cast<uint32_t>(code)
         );
         return false;
     }
@@ -305,7 +305,7 @@ bool TDengineConnector::execute(const StmtV2InsertData& data) {
             "{} execute failed: {} [code: 0x{:x}]",
             display_name_,
             taos_stmt2_error_(stmt_),
-            code
+            static_cast<uint32_t>(code)
         );
         return false;
     }
